+++ src/main/kotlin/com/booleworks/kjobs/api/persistence/redis/RedisConfig.kt
@@ -3,6 +3,9 @@
 
 package com.booleworks.kjobs.api.persistence.redis
 
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.days
+
 /**
  * Configuration for the [RedisDataPersistence].
  */
@@ -11,6 +14,13 @@ interface RedisConfig {
 
     val jobPattern: String get() = "job:*"
     val heartbeatPattern: String get() = "heartbeat:*"
+
+    /**
+     * The heartbeat expiration duration.
+     *
+     * A heartbeat has an expiration to prevent heartbeat entries of previous instances to remain forever in Redis.
+     */
+    val heartbeatExpiration: Duration get() = 1.days
 
     /**
      * The limit that is used for the Redis scan operations.